Allow redirecting to IDP specific websso route
The keystone global /websso route works well in the situation where
there is a SAML (or other) discovery page from which the user can select
the IDP they wish to use. However this doesn't allow for horizon to
choose a specific IDP that can be fetched without the discovery page.

To do this we need a way to specify the idp_id and the protocol of the
WEBSSO_CHOICES, so we add the WEBSSO_IDPS dictionary from which you can
correlate a CHOICE key with a idp_id, protocol pair.

Change-Id: I564955529077b55aded2792df8d300f6e4550be6